지도 수정, 날씨 수정

+++ summer_spot/views.py
@@ -12,7 +12,7 @@ from rest_framework.views import APIView
 from .serializer import Post_Serializer,Post_Categories_Serializer, Review_Serializer, Many_image_Serializer, \
     Post_list_Serializer
 from .models import Post
-from user_set.models import User, User_Categories
+from user_set.models import User
 from lion_project import settings
 import requests
 from datetime import datetime
@@ -24,12 +24,9 @@ class Post_make(APIView): # 휴양지 피드 만들기
         user_check = request.session.get('user')
         if user_check is None:
             return redirect('user_set:login')
-        if user_check is None:
-            return redirect('user_set:login')
         post = Post_Serializer()
-        post_cat = Post_Categories_Serializer()
         return Response(status=status.HTTP_200_OK, template_name='main/post_make.html',
-                        data={'post': post, 'post_cat': post_cat})
+                        data={'post': post})
 
     def post(self, request):
         user_check = request.session.get('user')
@@ -41,9 +38,13 @@ class Post_make(APIView): # 휴양지 피드 만들기
                 post.save()
                 return redirect('summer_spot:main_page')
         else:
-            error = str(post.errors)
+            va = post.errors
+            error = []
+            for key, value in va.items():
+                A = [key, value[0]]
+                error.append(A)
             return Response(status=status.HTTP_200_OK, template_name='main/post_make.html',
-                            data={'post': post, 'error': error})
+                            data={'post': post, "error": error})
 
 class Main_page(APIView): # 피드 리스트
     renderer_classes = [TemplateHTMLRenderer]
@@ -51,12 +52,9 @@ class Main_page(APIView): # 피드 리스트
         user_check = request.session.get('user')
         if user_check is None:
             return redirect('user_set:login')
-        user = User.objects.get(id=user_check)
-        user_re = User_Categories.objects.filter(user=user_check)
-        post_re = Post.objects.all()
+        main_table_values = User.objects.filter(id=user_check).values('mood', 'personnel', 'leisure', 'rental_item')
+        post_table_values = Post.objects.all().values('mood', 'personnel', 'leisure', 'rental_item', 'id')
 
-        main_table_values = user_re.values()
-        post_table_values = post_re.values()
         if post_table_values:
             first_queryset = main_table_values
             second_queryset = post_table_values
@@ -82,7 +80,7 @@ class Main_page(APIView): # 피드 리스트
             return Response(status=status.HTTP_200_OK, template_name='lion_2/index.html', data={"post": posts})
 
         else:
-            return Response(status=status.HTTP_200_OK, template_name='lion_2/index.html', data={"post": post_re.values()})
+            return Response(status=status.HTTP_200_OK, template_name='lion_2/index.html', data={"post": post_table_values})
 
 class Post_region(APIView): # 지역 검색
     renderer_classes = [TemplateHTMLRenderer]
@@ -90,12 +88,9 @@ class Post_region(APIView): # 지역 검색
         user_check = request.session.get('user')
         if user_check is None:
             return redirect('user_set:login')
-        post_re = Post.objects.filter(region=request.data['searchbox'])
-        user = User.objects.get(id=user_check)
-        user_re = User_Categories.objects.filter(user=user)
+        main_table_values = User.objects.filter(id=user_check).values('mood', 'personnel', 'leisure', 'rental_item')
+        post_table_values = Post.objects.filter(region=request.data['searchbox']).values('mood', 'personnel', 'leisure', 'rental_item', 'id')
 
-        main_table_values = user_re.values()
-        post_table_values = post_re.values()
         if post_table_values:
             first_queryset = main_table_values
             second_queryset = post_table_values
@@ -120,17 +115,20 @@ class Post_region(APIView): # 지역 검색
                 return Response(status=status.HTTP_200_OK, template_name='lion_2/index.html', data={"post": posts})
 
         else:
-                return Response(status=status.HTTP_200_OK, template_name='lion_2/index.html', data={"post": post_re.values()})
+                return Response(status=status.HTTP_200_OK, template_name='lion_2/index.html', data={"post": post_table_values})
 class Post_search(APIView): # 피드 검색
     renderer_classes = [TemplateHTMLRenderer]
+    def get(self, request):
+        user_check = request.session.get('user')
+        if user_check is None:
+            return redirect('user_set:login')
+        return Response(status=status.HTTP_200_OK, template_name='lion_2/keyword.html')
+
     def post(self, request):
         user_check = request.session.get('user')
-        user = User.objects.get(id=user_check)
-        user_re = User_Categories.objects.filter(user=user_check)
-        post_re = Post.objects.all()
+        main_table_values = User.objects.filter(id=user_check).values('mood', 'personnel', 'leisure', 'rental_item')
+        post_table_values = Post.objects.all().values('mood', 'personnel', 'leisure', 'rental_item', 'id')
 
-        main_table_values = user_re.values()
-        post_table_values = post_re.values()
         if post_table_values:
             first_queryset = main_table_values
             second_queryset = post_table_values
@@ -147,20 +145,28 @@ class Post_search(APIView): # 피드 검색
             posts = []
             filters = Q()  # 빈 Q 객체 생성
 
-            if request.data['ambience']:
-                filters &= Q(ambience__icontains=request.data['ambience'])
+            if 'mood' in request.data:
+                filters &= Q(mood__in=request.data.getlist('mood'))
 
-            if request.data['personnel']:
-                filters &= Q(personnel__icontains=request.data['personnel'])
+            if 'personnel' in request.data:
+                filters &= Q(personnel__in=request.data.getlist('personnel'))
 
-            if request.data['view']:
-                filters &= Q(view__icontains=request.data['view'])
+            if 'leisure' in request.data:
+                filters &= Q(leisure__in=request.data.getlist('leisure'))
 
-            if request.data['good_place']:
-                filters &= Q(good_place__icontains=request.data['good_place'])
+            if 'rental_item' in request.data:
+                filters &= Q(rental_item__in=request.data.getlist('rental_item'))
 
-            if request.data['rental_item']:
-                filters &= Q(rental_item__icontains=request.data['rental_item'])
+
+            mood_choices = dict(Post.MOOD)
+            personnel_choices = dict(Post.PERSONNEL)
+            leisure_choices = dict(Post.LEISURE)
+            rental_item_choices = dict(Post.RENTAL_ITEM)
+            mood_key = [mood_choices[key] for key in request.data.getlist('mood')]
+            personnel_key = [personnel_choices[key] for key in request.data.getlist('personnel')]
+            leisure_key = [leisure_choices[key] for key in request.data.getlist('leisure')]
+            rental_item_key = [rental_item_choices[key] for key in request.data.getlist('rental_item')]
+
             for grouped_mapping in grouped_queries:
                 filter_condition = Q()
                 for fields_mapping in grouped_mapping:
@@ -169,13 +175,16 @@ class Post_search(APIView): # 피드 검색
                 my_objects = Post.objects.filter(filter_condition)
                 posts.append(my_objects.order_by('-user_check').filter(filters))
 
-            print(posts)
 
-            return Response(status=status.HTTP_200_OK, template_name='main/main_page.html', data={"post": posts})
+            return Response(status=status.HTTP_200_OK, template_name='main/main_page.html', data={"post": posts,
+                                                                                                  "mood_key": mood_key,
+                                                                                                  'personnel_key': personnel_key,
+                                                                                                  'leisure_key': leisure_key,
+                                                                                                  'rental_item_key': rental_item_key})
 
         else:
             return Response(status=status.HTTP_200_OK, template_name='main/main_page.html',
-                            data={"post": post_re.values()})
+                            data={"post": post_table_values})
 
 
 class Select_post(APIView): # 선택한 피드
@@ -188,9 +197,13 @@ class Select_post(APIView): # 선택한 피드
         post.user_check += 1
         post.save()
         review_set = Review_Serializer(data=request.data)
-        return Response(status=status.HTTP_200_OK, template_name='main/post.html', data={"post": post, 'pk': pk,
+        count = []
+        for i in range(len(post.post_image.all())):
+            count.append(i+1)
+        return Response(status=status.HTTP_200_OK, template_name='lion_2/detail.html', data={"post": post, 'pk': pk,
                                                                                          'review': review_set,
-                                                                                         'check_user': check_user})
+                                                                                         'check_user': check_user,
+                                                                                             "count": count})
 
 class Weather(APIView): # 날씨
     renderer_classes = [TemplateHTMLRenderer]
@@ -228,6 +241,8 @@ class Weather(APIView): # 날씨
             weather = 'cloudy-windy'
         elif weather == 'mist' and 'smoke' and 'haze' and 'dust':
             weather = 'fog'
+        elif weather == 'clouds':
+            weather = 'cloudy'
         now_icon = now
         if now > 12:
             now_icon = now-12
@@ -290,15 +305,23 @@ class Post_retouch(APIView): # 피드 수정
         post_check = Post_Serializer(post, data=request.data)
         if post_check.is_valid():
             post_check.save()
-            return redirect('summer_spot:main_page')
+            return redirect('summer_spot:post', pk)
         else:
+            va = post_check.errors
+            error = []
+            for key, value in va.items():
+                A = [key, value[0]]
+                error.append(A)
             return Response(status=status.HTTP_200_OK, template_name='main/post_retouch.html',
-                                data={"post": post_check, 'pk': pk})
+                                data={"post": post_check, 'pk': pk, 'error': error})
 
 class Post_images_del(APIView): # 피드 이미지 삭제
     renderer_classes = [TemplateHTMLRenderer]
 
     def get(self, request, pk, pk2):
+        user_check = request.session.get('user')
+        if user_check is None:
+            return redirect('user_set:login')
         post = Post.objects.get(id=pk)
         post.post_image.get(id=pk2).delete()
         return redirect('summer_spot:post', pk)
@@ -306,18 +329,27 @@ class Post_images_del(APIView): # 피드 이미지 삭제
 class Review_del(APIView): # 피드 댓글 삭제
 
     def get(self, request, pk, pk2):
+        user_check = request.session.get('user')
+        if user_check is None:
+            return redirect('user_set:login')
         post = Post.objects.get(id=pk)
         post.post_review.get(id=pk2).delete()
         return redirect('summer_spot:post', pk)
 
 class Post_del(APIView): # 포스트 삭제
     def get(self, request, pk):
+        user_check = request.session.get('user')
+        if user_check is None:
+            return redirect('user_set:login')
         Post.objects.get(id=pk).delete()
         return redirect('summer_spot:main_page')
 
-class Post_list_make(APIView): # 포스트 리스트 삭제
+class Post_list_make(APIView): # 포스트 리스트 만들기
     renderer_classes = [TemplateHTMLRenderer]
     def get(self, request, pk):
+        user_check = request.session.get('user')
+        if user_check is None:
+            return redirect('user_set:login')
         form = Post_list_Serializer()
         return Response(status=status.HTTP_200_OK, template_name='main/post_list_make.html',
                         data={"form": form, 'pk': pk})
@@ -331,12 +363,20 @@ class Post_list_make(APIView): # 포스트 리스트 삭제
             post.save()
             return redirect('summer_spot:post', pk)
         else:
+            va = form.errors
+            error = []
+            for key, value in va.items():
+                A = [key, value[0]]
+                error.append(A)
             return Response(status=status.HTTP_200_OK, template_name='main/post_list_make.html',
-                            data={"form": form, 'pk': pk})
+                            data={"form": form, 'pk': pk, 'error':error})
 class Post_list_retouch(APIView): # 피드 리스트 수정
     renderer_classes = [TemplateHTMLRenderer]
 
     def get(self, request, pk, pk2):
+        user_check = request.session.get('user')
+        if user_check is None:
+            return redirect('user_set:login')
         post = Post.objects.get(id=pk)
         post_list = post.post_list.get(id=pk2)
         post_check = Post_list_Serializer(instance=post_list)
@@ -350,26 +390,44 @@ class Post_list_retouch(APIView): # 피드 리스트 수정
         post_list_check = Post_list_Serializer(post_list, data=request.data)
         if post_list_check.is_valid():
             post_list_check.save()
-            print("asdasdas")
             return redirect('summer_spot:post_list', pk, pk2)
         else:
+            va = post_list_check.errors
+            error = []
+            for key, value in va.items():
+                A = [key, value[0]]
+                error.append(A)
             return Response(status=status.HTTP_200_OK, template_name='main/post_list_retouch.html',
-                            data={"form": post_list_check, 'pk': pk, 'pk2': pk2})
+                            data={"form": post_list_check, 'pk': pk, 'pk2': pk2, 'error': error})
 
 class Post_list(APIView): # 선택한 피드 리스트
     renderer_classes = [TemplateHTMLRenderer]
     def get(self, request, pk, pk2):
+        user_check = request.session.get('user')
+        if user_check is None:
+            return redirect('user_set:login')
         post = Post.objects.get(id=pk)
         post_list = post.post_list.get(id=pk2)
         # output_text = post_list.list_content.replace('\n', '<br>')
         # output_text = html.escape(output_text)
         return Response(status=status.HTTP_200_OK, template_name='main/post_list.html',
-                        data={"post_list": post_list, 'pk': pk, 'pk2': pk2})
+                        data={"post_list": post_list, 'pk': pk, 'pk2': pk2, 'user_check':user_check})
 
 class Post_list_del(APIView): # 선택한 피드 리스트 삭제
 
     def get(self, request, pk, pk2):
+        user_check = request.session.get('user')
+        if user_check is None:
+            return redirect('user_set:login')
         post = Post.objects.get(id=pk)
         post.post_list.get(id=pk2).delete()
         post.save()
         return redirect('summer_spot:post', pk)
+
+class Weather_search(APIView):
+    renderer_classes = [TemplateHTMLRenderer]
+    def get(self, request):
+        user_check = request.session.get('user')
+        if user_check is None:
+            return redirect('user_set:login')
+        return Response(status=status.HTTP_200_OK, template_name='lion_2/weather.html')
